Refactor to e2e test to keep shared state
